pyproject.toml
[project]
name = "personal-account-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.220.0",
    "bedrock-agentcore-starter-toolkit>=0.1.21",
    "constructs>=10.0.0",
    "gitpython>=3.1.45",
    "strands-agents>=1.13.0",
    "strands-agents-tools>=0.2.12",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.14.1",
]
